Reference the DID 1.0 Syntax for the definition and syntax for DID methods.
This changes from referencing the DID Registries Note, which does not define what DID methods are or how they are formatted.

In addition, the syntax currently in use (e.g. `did:example:`) appears to be a local invention, so we now describe the production to form this string from a given DID method.